Update README with information on using parquet as tempformat
